fix(bundle1): Copilot cloud-review on heartbeat + runbook (#532)
Copilot R1 found 6 issues on Bundle 1. Applied 3, skipped 3 duplicates.

M-1 (agents/heartbeat/README.md): Healthcheck section said success
marker is written 'after a confirmed heartbeat transaction with a
successful webhook POST', but RunnerCastHeartbeat.callHeartbeat()
treats webhook POST as best-effort and always calls the writer after
the transaction. Fixed the doc to match actual behavior — marker is
written after confirmed heartbeat progress, webhook is best-effort.

M-2 (packages/runner/src/heartbeatScheduler.ts): scheduler imported
writeHeartbeatSuccessFile from runnerCastHeartbeat.ts, which pulled in
the Viem client stack as a transitive dep for any consumer of just the
scheduler. Extracted writeHeartbeatSuccessFile into a new pure-IO
module heartbeatSuccessFile.ts. Scheduler now imports from there.
runnerCastHeartbeat.ts re-exports for backward-compat. Updated test
to import from the new module + match new warning prefix.

L-1 (docs/runbooks/anvil-fork-dev-rpc.md): Referenced
'make -C agents reset-anvil PROFILE=dev' but agents/Makefile is Bundle 3
(PR #548), not present in Bundle 1. Fixed to use root 'make reset-anvil
PROFILE=dev' with a note that the per-elder Makefile invocation becomes
available after Bundle 3 merges.

Skipped: 3 Convex 1.17.4 vs 1.39.1 docs findings — already documented
in PR #547 Bundle 3 runbook + issue #531.

Validation:
- git diff --check: passed
- Scheduler grep for runnerCastHeartbeat: clean
- agents/Makefile reference grep: only the explanatory note remains
- pnpm typecheck not runnable locally (no node_modules)
