+++ agents/heartbeat/README.md
@@ -52,8 +52,8 @@ compose `anvil-fork` service and viem's `baseSepolia` chain config.
 
 The compose healthcheck reads `/tmp/last-heartbeat-success` (path configurable
 via `HEARTBEAT_SUCCESS_FILE_OVERRIDE`). The runner writes the current Unix timestamp
-there after either (a) a confirmed heartbeat transaction with a successful
-webhook POST, OR (b) the receipt-timeout recovery path when on-chain
+there after either (a) a confirmed heartbeat transaction, OR (b) the
+receipt-timeout recovery path when on-chain
 `nextHeartbeatAtTs` is observed to have advanced (no confirmed receipt, no
 webhook POST). The marker indicates the scheduler is making forward progress,
 NOT that the most recent transaction was confirmed end-to-end. A timestamp
