+++ packages/runner/src/heartbeatScheduler.ts
@@ -1,6 +1,6 @@
 import { HeartbeatRateLimitedError, type IHeartbeatCaller } from '@clan-world/agents/seams';
 import type { IConvexClient, RunnerStatusUpdate } from '@clan-world/shared/adapters';
-import { writeHeartbeatSuccessFile } from './runnerCastHeartbeat';
+import { writeHeartbeatSuccessFile } from './heartbeatSuccessFile';
 import { sendTelegramAlert } from './telegramAlert';
 
 export type HeartbeatFireResult = RunnerStatusUpdate['lastFireResult'];
